+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## unreleased
 
+ - Support parsing of SMT-LIB FPA Theory literals and add the theory `smt.float`
+   that enables thte support for the SMT-LIB FPA, and that can be used with the
+   cmdline option `--enable-theory smt.float` (#1358)
  - Bump minimal cmdliner version to 2.0 (#1340)
 
 ## v2.6.2